m and n can be in either order for D, FR, K, P, PW, T, and X commands. Revised inline scan functions to act as no-ops if not in exec mode. Fixed off-by-one error in F2 definition in vtedit.tec.

+++ src/file.c
@@ -210,7 +210,7 @@ char *init_filename(const char *name, uint_t len, bool colon)
     // Next, scan for match control characters.
 
     name = build_trimmed(name, (uint_t)(uint)nbytes);
-    len = strlen(name);
+    len = (uint)strlen(name);
 
     // Ensure that there are no non-graphic ASCII characters in the filespec.
